Define new function islandora_sort_ordered
This function can replace calls to uasort with drupal_sort_weight. This
function will return a consistant sort order between PHP5 and PHP7 for
elements of the same weight by preserving the order of elements with the
same weight.

Under both PHP5 and PHP7 the order of elements with the same weight are
not guarenteed, and in practice the order of elements returned is
different between PHP5 and PHP7. This could cause some issues where the
order is changed depending on PHP version. This function gives us the
option of a consistant sort order.

This pull request also defines some new unit tests for the new function.